+++ api/src/main/java/com/staxpayments/api/repository/CustomerRepository.kt
@@ -1,0 +1,7 @@
+package com.staxpayments.api.repository
+
+import com.staxpayments.api.models.Customer
+
+interface CustomerRepository {
+   suspend fun customer(): Customer
+}
